Create Quiz.java
This version of Quiz.java has been modified according to the instructions given in the assignment. It now includes the ask and check methods, and asks the user three questions. It also keeps track of the number of questions asked and the number of correct answers given, displaying the score at the end.